requirements_base.txt
# local package
-e .

# external requirements
click
joblib
tqdm
numpy
black
flake8
python-dotenv
pytest
openai
python-telegram-bot[webhooks]
datadog_api_client